MDE/PKFE-7 component refactoring and added jsdocs
