--- app/front-end/src/features/editor/components/toolbar.tsx
@@ -1,3 +1,0 @@
-export const Toolbar: React.FC = () => {
-  return <>{/* Toolbar */}</>;
-};
