+++ app/front-end/src/features/editor/components/toolbarView/index.ts
@@ -1,5 +1,9 @@
-export { ToolbarGroup, ToolbarGroupItem } from './toolbarGroup';
-export type { ToolbarGroupItemProps, ToolbarGroupProps } from './toolbarGroup';
-export { ToolbarGroupsSelector, ToolbarGroupsSelectorItem } from './toolbarGroupsSelector';
-export type { ToolbarGroupsSelectorItemProps, ToolbarGroupsSelectorProps } from './toolbarGroupsSelector';
+export { ToolbarGroup } from './toolbarGroup';
+export type { ToolbarGroupProps } from './toolbarGroup';
+export { ToolbarGroupItem } from './toolbarGroupItem';
+export type { ToolbarGroupItemProps } from './toolbarGroupItem';
+export { ToolbarGroupsSelector } from './toolbarGroupsSelector';
+export type { ToolbarGroupsSelectorProps } from './toolbarGroupsSelector';
+export { ToolbarGroupsSelectorItem } from './toolbarGroupsSelectorItem';
+export type { ToolbarGroupsSelectorItemProps } from './toolbarGroupsSelectorItem';
 export { ToolbarView } from './toolbarView';
