+++ app/front-end/src/features/editor/components/toolbarView/toolbarGroupsSelector.tsx
@@ -1,45 +1,30 @@
-import { Button, List, useTheme } from '@mui/material';
+import { List } from '@mui/material';
 
 export interface ToolbarGroupsSelectorProps {
   children: React.ReactNode;
 }
 
+/**
+ * ToolbarGroupsSelector component renders a list container that houses toolbar group items.
+ *
+ * @description This component is a wrapper around the Material-UI `List` component. It is used to display
+ * a list of toolbar group items within a scrollable area. The height of the list is fixed at 25%, and
+ * it has zero padding with an automatic overflow handling.
+ *
+ * @component
+ *
+ * @example
+ * // Example usage of the ToolbarGroupsSelector component
+ * return (
+ *   <ToolbarGroupsSelector>
+ *     <ToolbarGroupsSelectorItem id="1" label="Group 1" onClick={handleClick} />
+ *     <ToolbarGroupsSelectorItem id="2" label="Group 2" onClick={handleClick} />
+ *   </ToolbarGroupsSelector>
+ * );
+ *
+ * @param {React.ReactNode} children - The child elements to be rendered inside the list.
+ * @returns {JSX.Element} The rendered List component.
+ */
 export const ToolbarGroupsSelector: React.FC<ToolbarGroupsSelectorProps> = ({ children }) => {
   return <List sx={{ height: '25%', p: '0', overflow: 'auto' }}>{children}</List>;
-};
-
-export interface ToolbarGroupsSelectorItemProps {
-  id: string;
-  label: string;
-  onClick: () => void;
-  groupRef?: string;
-}
-
-export const ToolbarGroupsSelectorItem: React.FC<ToolbarGroupsSelectorItemProps> = ({
-  id,
-  label,
-  onClick,
-  groupRef,
-}) => {
-  const Theme = useTheme();
-
-  return (
-    <Button
-      id={id}
-      sx={{
-        height: '100%',
-        bgcolor: groupRef === id ? Theme.palette.background.paper : Theme.palette.action.selected,
-        borderRadius: '0.625rem 0.625rem 0rem 0rem',
-        px: '3rem',
-        fontWeight: 'bold',
-        color: Theme.palette.text.primary,
-        ':hover': {
-          backgroundColor: Theme.palette.background.paper,
-        },
-      }}
-      onClick={() => onClick()}
-    >
-      {label}
-    </Button>
-  );
 };
